+++ src/Tablist.js
@@ -58,6 +58,8 @@ import {
   EVENT_MOUSEDOWN,
   HIDDEN_CLASS,
   SHOWN_CLASS,
+  UNTIL_FOUND,
+  MODE_HIDDEN_UNTIL_FOUND,
 } from "./helpers/constants";
 import Base from "./helpers/Base.js";
 import {
@@ -118,7 +120,15 @@ const OPTION_STATE_ATTRIBUTE = "stateAttribute";
 const OPTION_ALWAYS_EXPANDED = "alwaysExpanded";
 const OPTION_MULTI_EXPAND = "multiExpand";
 
-const TABLIST_SECONDARY_METHODS = ["getTab", "isTab", "initTab", "initTabs"];
+const ACTION_INIT_TAB = "initTab";
+const ACTION_DESTROY_TAB = "destroyTab";
+
+const TABLIST_SECONDARY_METHODS = [
+  "getTab",
+  "isTab",
+  ACTION_INIT_TAB,
+  "initTabs",
+];
 
 const DEFAULT_ACCORDION_OPTIONS = {
   [OPTION_ALWAYS_EXPANDED]: false,
@@ -338,7 +348,7 @@ class Tablist extends Base {
   initTab(tab) {
     if (this.getTab(tab)) return;
 
-    const { tabs, tablist, opts, shownTabs, on, off } = this;
+    const { tabs, tablist, opts, shownTabs, on, off, emit } = this;
     const index = tabs.length;
 
     const item = isString(opts[ITEM])
@@ -361,6 +371,10 @@ class Tablist extends Base {
     }
 
     if (!tabpanel) return;
+
+    if (opts[HIDE_MODE] === HIDDEN && tabpanel[HIDDEN] === UNTIL_FOUND) {
+      opts[HIDE_MODE] = MODE_HIDDEN_UNTIL_FOUND;
+    }
 
     const uuid = uuidGenerator();
     const id = (tabpanel.id ||= uuid);
@@ -385,7 +399,7 @@ class Tablist extends Base {
     });
     on(tab, EVENT_CLICK, (event) => {
       event.preventDefault();
-      this.toggle(event.currentTarget, null, { event, trigger: tab });
+      this.toggle(tab, null, { event, trigger: tab });
     });
 
     const destroy = ({
@@ -436,6 +450,8 @@ class Tablist extends Base {
           removeAttribute(elem, HIDDEN, INERT);
         }
       });
+
+      emit(ACTION_DESTROY_TAB, tabInstance);
     };
 
     const toggleDisabled = (s = null) => {
@@ -456,7 +472,7 @@ class Tablist extends Base {
       return !disabled;
     };
 
-    const elems = [tab, item, tabpanel];
+    const elems = [tab, item, tabpanel].filter(Boolean);
     const tabInstance = {
       id,
       uuid,
@@ -495,6 +511,9 @@ class Tablist extends Base {
     }
 
     tabs.push(tabInstance);
+
+    emit(ACTION_INIT_TAB, tabInstance);
+
     return tabInstance;
   }
   isTab(tab, value) {