+++ src/helpers/constants/index.js
@@ -112,6 +112,7 @@ export const FALSE = "false";
 
 export const doc = document;
 export const body = doc.body;
+export const ROOT_ELEM = doc.documentElement;
 export const BODY = "body";
 
 export const ENTER = "enter";
@@ -161,13 +162,13 @@ export const POPOVER_API_MODE_MANUAL = "manual";
 export const PLACEHOLDER = "placeholder";
 
 export const EVENT_INIT = ACTION_INIT;
-export const EVENT_BEFORE_INIT = BEFORE + upperFirst(EVENT_INIT);
+export const EVENT_BEFORE_INIT = BEFORE + "Init";
 export const EVENT_DESTROY = ACTION_DESTROY;
-export const EVENT_BEFORE_DESTROY = BEFORE + upperFirst(EVENT_DESTROY);
-export const EVENT_BEFORE_SHOW = BEFORE + upperFirst(SHOW);
+export const EVENT_BEFORE_DESTROY = BEFORE + "Destroy";
+export const EVENT_BEFORE_SHOW = BEFORE + "Show";
 export const EVENT_SHOW = SHOW;
 export const EVENT_SHOWN = SHOWN;
-export const EVENT_BEFORE_HIDE = BEFORE + upperFirst(HIDE);
+export const EVENT_BEFORE_HIDE = BEFORE + "Hide";
 export const EVENT_HIDE = HIDE;
 export const EVENT_HIDDEN = HIDDEN;
 export const EVENT_PAUSE = ACTION_PAUSE;
@@ -195,6 +196,10 @@ export const EVENT_SCROLL = SCROLL;
 export const EVENT_RESIZE = "resize";
 export const EVENT_DISABLED = DISABLED;
 export const EVENT_ENABLED = "enabled";
+export const EVENT_BEFORE_MATCH = BEFORE + "match";
+
+export const UNTIL_FOUND = "until-found";
+export const MODE_HIDDEN_UNTIL_FOUND = HIDDEN + "-" + UNTIL_FOUND;
 
 export const ARIA = "aria";
 export const ARIA_CONTROLS = ARIA + "-controls";