+++ src/helpers/utils/isShown.js
@@ -2,7 +2,9 @@ import {
   ACTION_REMOVE,
   CLASS_HIDDEN_MODE,
   CLASS_SHOWN_MODE,
+  HIDDEN,
   HIDDEN_CLASS,
+  MODE_HIDDEN_UNTIL_FOUND,
   SHOWN_CLASS,
 } from "../constants/index.js";
 import { inDOM } from "../dom/index.js";
@@ -14,5 +16,7 @@ export default (elem, hideMode) => {
     ? !elem.classList.contains(HIDDEN_CLASS)
     : hideMode === CLASS_SHOWN_MODE
     ? elem.classList.contains(SHOWN_CLASS)
-    : !elem.hasAttribute(hideMode);
+    : !elem.hasAttribute(
+        hideMode === MODE_HIDDEN_UNTIL_FOUND ? HIDDEN : hideMode,
+      );
 };